DRY my code.
I submitted my guessing-game code to Exercism, and the analysis suggested
a default argument. That resulted in me learning about bodyless functions,
AKA function headers. The code is DRYer now. Cool.

I believe that bodyless functions are going to be taught in a future exercise
that I haven't yet unlocked. I got enough of the idea for now to DRY my code.